pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "batframework"
version = "0.1.13"
authors = [
    { name="Turan Baturay", email="baturayturan@gmail.com" }
]
description = "Pygame framework for making games easier."
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies=[
    "pygame-ce"
]

[project.urls]
"Homepage" = "https://github.com/TuranBaturay/batFramework"
